Finally fixed config creation issue
